Completing required funtionality

+++ src/main/java/com/unosquare/demo/accountmanagement/util/Constants.java
@@ -5,15 +5,15 @@ public final class Constants {
 
 	public static final String EMPTY_STRING = "";
 	
-	public static final long MAX_PIN_VALUE = 9999;
-	public static final long MIN_PIN_VALUE = 1000;
-	public static final long MAX_ACCOUNT_VALUE = 999999999999l;
-	public static final long MIN_ACCOUNT_VALUE = 100000000000l;
+	public static final long MAX_PIN_VALUE = 9999L;
+	public static final long MIN_PIN_VALUE = 1000L;
+	public static final long MAX_ACCOUNT_VALUE = 999999999999L;
+	public static final long MIN_ACCOUNT_VALUE = 100000000000L;
 	
-	public static final int TRASACTION_TYPE_DEPOSIT = 1;
-	public static final int TRASACTION_TYPE_WITHDRAWAL = 2;
-	public static final int TRASACTION_TYPE_DEBIT = 3;
-	public static final int TRASACTION_TYPE_CHECK = 4;
+	public static final Long TRASACTION_TYPE_DEPOSIT = 1L;
+	public static final Long TRASACTION_TYPE_WITHDRAWAL = 2L;
+	public static final Long TRASACTION_TYPE_DEBIT = 3L;
+	public static final Long TRASACTION_TYPE_CHECK = 4L;
 	
 	public static final int ERROR_CODE = -1;
 	public static final int SUCCESS_CODE = 1;
